vL2.11: ELiminació en cascada que funciona correctament si es la última instancia

+++ web/views.py
@@ -1,4 +1,5 @@
 from django.core.exceptions import PermissionDenied
+from django.db import transaction
 from django.http import JsonResponse, HttpResponseRedirect
 from django.shortcuts import render, redirect, get_object_or_404
 from django.urls import reverse_lazy, reverse
@@ -50,12 +51,32 @@ class DeletePlaylist(LoginRequiredMixin,CheckIsOwnerMixin, DeleteView):
     template_name = 'playlist_confirm_delete.html'  # Plantilla para confirmar la eliminación
     success_url = reverse_lazy('web:playlist_list')  # URL a la que se redireccionará después de la eliminación
 
+    def delete(self, request, *args, **kwargs):
+        playlist = self.get_object()
 
+        # Obtener todas las canciones asociadas a la playlist
+        songs = playlist.songs.all()
+
+        # Eliminar las canciones asociadas solo a esta playlist
+        for song in songs:
+            if playlist in song.songs_to_playlist.all():
+                song.delete()
+
+        # Luego de eliminar las canciones asociadas, eliminar la playlist
+        response = super().delete(request, *args, **kwargs)
+
+        # Verificar si alguna de las canciones eliminadas no está referenciada por ninguna otra playlist
+        for song in songs:
+            if not Playlist.objects.filter(songs=song).exists():
+                song.delete()
+
+        return response
 
 class PlaylistCreate(LoginRequiredMixin, CreateView):
     model = Playlist
     template_name = 'form.html'
     form_class = PlaylistForm
+
 
     def form_valid(self, form):
         form.instance.user = self.request.user  # Assignem l'usuari actual a la llista de reproducció
@@ -76,7 +97,6 @@ class SongCreate(LoginRequiredMixin,CheckIsOwnerMixin, CreateView):
 
     def post(self, request, *args, **kwargs):
         if request.method == 'POST':
-
             # Obtener los datos de la solicitud POST
             nombre_cancion = request.POST.get('nombre_cancion')
             nombre_artista = request.POST.get('nombre_artista')
@@ -93,24 +113,28 @@ class SongCreate(LoginRequiredMixin,CheckIsOwnerMixin, CreateView):
             else:
                 duracion_timedelta = None
 
-            # Obtener o crear el artista
-            artista, _ = Artist.objects.get_or_create(name=nombre_artista)
+            with transaction.atomic():
+                # Verificar si la canción ya existe
+                cancion_existente = Song.objects.filter(title=nombre_cancion, artists__name=nombre_artista).first()
 
-            # Crear la canción y asociar el artista
-            cancion = Song.objects.create(title=nombre_cancion, album=nombre_album, duration=duracion_timedelta, url_imagen=url_imagen, url_cancion=url_cancion)
-            cancion.artists.add(artista)
-
-            # Obtener la playlist y agregar la canción
-            playlist_pk = self.kwargs['pk']
-            playlist = Playlist.objects.get(pk=playlist_pk)
-            playlist.songs.add(cancion)
-
-            # Devolver una respuesta JSON
-            return JsonResponse({'mensaje': 'Canción agregada a la playlist correctamente'})
+                if cancion_existente:
+                    # Si la canción ya existe, no es necesario crear una nueva
+                    playlist_pk = self.kwargs['pk']
+                    playlist = Playlist.objects.get(pk=playlist_pk)
+                    playlist.songs.add(cancion_existente)
+                    return JsonResponse({'mensaje': 'Canción ya existente agregada a la playlist correctamente'})
+                else:
+                    # Si la canción no existe, crearla y asociarla al artista
+                    artista, _ = Artist.objects.get_or_create(name=nombre_artista)
+                    cancion = Song.objects.create(title=nombre_cancion, album=nombre_album, duration=duracion_timedelta,
+                                                  url_imagen=url_imagen, url_cancion=url_cancion)
+                    cancion.artists.add(artista)
+                    playlist_pk = self.kwargs['pk']
+                    playlist = Playlist.objects.get(pk=playlist_pk)
+                    playlist.songs.add(cancion)
+                    return JsonResponse({'mensaje': 'Canción agregada a la playlist correctamente'})
         else:
             return JsonResponse({'error': 'Método no permitido'}, status=405)
-
-
 class SongDelete( LoginRequiredMixin, CheckIsOwnerMixin, DeleteView):
     model = Song
 
@@ -119,11 +143,19 @@ class SongDelete( LoginRequiredMixin, CheckIsOwnerMixin, DeleteView):
         playlist = get_object_or_404(Playlist, pk=self.kwargs['pk'])
         # Obtener la instancia de la canción que se desea eliminar
         song = get_object_or_404(Song, pk=self.kwargs['pkr'])
-        # Eliminar la canción de la lista de reproducción
+
+        # Verificar si hay otras listas de reproducción que contienen esta canción
+        other_playlists_with_song = Playlist.objects.filter(songs=song).exclude(pk=playlist.pk)
+
+        # Si no hay otras listas de reproducción que contienen esta canción, eliminarla
+        if not other_playlists_with_song.exists():
+            song.delete()
+
+        # Eliminar la canción de la lista de reproducción actual
         playlist.songs.remove(song)
-        # Guardar los cambios
+        # Guardar los cambios en la lista de reproducción
         playlist.save()
-        song.delete()
+
         # Redirigir a la página de detalles de la lista de reproducción
         return redirect('web:playlist_detail', pk=playlist.id)
